+++ versiondb/client/restore_app_db.go
@@ -283,7 +283,7 @@ func encodeNode(w io.Writer, node memiavl.PersistedNode) error {
 	if _, err := w.Write(buf[:n]); err != nil {
 		return err
 	}
-	n = binary.PutVarint(buf[:], node.Version())
+	n = binary.PutVarint(buf[:], int64(node.Version()))
 	if _, err := w.Write(buf[:n]); err != nil {
 		return err
 	}