+++ CHANGELOG.md
@@ -20,6 +20,10 @@
 
 - [#833](https://github.com/crypto-org-chain/cronos/pull/833) Fix rollback command.
 
+### Improvements
+
+- [#890](https://github.com/crypto-org-chain/cronos/pull/890) optimize memiavl snapshot format.
+
 *Feb 09, 2022*
 
 ## v1.0.4